split euler stuff to separate module
